ADDED INTERNAL TRANSFER BETWEEN USERS
Added the option to transfer money between users of the same bank.

Added the lookup account number to get account name feature in the transfer form.

Fixed the transfer form to show the account name of the account to transfer to.

Fixed the transferController to handle the transfer between users.

Added API routes to lookup account number to get account name.

Added routes to the web.php file to handle the transfer form.

Created the AccountController to handle Account changes.